+++ example/http_server.ml
@@ -30,7 +30,7 @@ let text =
    the well, and noticed that they were filled with cupboards......"
 ;;
 
-let handler ctx _request = return (Server.respond_string ctx text)
+let handler _request = return (Response.create ~body:(Body.string text) `Ok)
 
 let run sock =
   let server =
@@ -42,7 +42,7 @@ let run sock =
       (Tcp.Where_to_listen.of_port sock)
       (fun _addr reader writer ->
       let server = Shuttle_http.Server.create reader writer in
-      Server.run server (handler server))
+      Server.run server handler)
   in
   Log.Global.info
     !"Server listening on: %s"