+++ CHANGES.md
@@ -1,6 +1,7 @@
 # next
 
 * Forward ssl connection details to shuttle_ssl server handlers
+* Remove listen and with_connection from shuttle_ssl. Use upgrade_server_connection and upgrade_client_connection with the regular Connection module provided within Shuttle.
 
 # 0.8.1